refactor to dedup this function
