+++ crates/archive/src/tarball.rs
@@ -1,12 +1,13 @@
 //! Provides types and functions for fetching and unpacking a Node installation
 //! tarball in Unix operating systems.
 
-use std::fs::{self, File};
+use std::fs::File;
 use std::io::{Read, Seek, SeekFrom};
 use std::path::Path;
 
 use failure::{self, Fail};
 use flate2::read::GzDecoder;
+use fs_utils::ensure_containing_dir_exists;
 use headers_011::Headers011;
 use progress_read::ProgressRead;
 use reqwest;
@@ -28,40 +29,6 @@ pub struct Tarball {
     uncompressed_size: Option<u64>,
     data: Box<Read>,
     origin: Origin,
-}
-
-/// Thrown when the containing directory could not be determined
-#[derive(Fail, Debug)]
-#[fail(display = "Could not determine directory information for {}", path)]
-struct ContainingDirError {
-    path: String,
-}
-
-/// Thrown when the containing directory could not be determined
-#[derive(Fail, Debug)]
-#[fail(display = "Could not create directory {}", dir)]
-struct CreateDirError {
-    dir: String,
-}
-
-/// This creates the parent directory of the input path, assuming the input path is a file.
-pub fn ensure_containing_dir_exists<P: AsRef<Path>>(path: &P) -> Result<(), failure::Error> {
-    path.as_ref()
-        .parent()
-        .ok_or(
-            ContainingDirError {
-                path: path.as_ref().to_string_lossy().to_string(),
-            }
-            .into(),
-        )
-        .and_then(|dir| {
-            fs::create_dir_all(dir).map_err(|_| {
-                CreateDirError {
-                    dir: dir.to_string_lossy().to_string(),
-                }
-                .into()
-            })
-        })
 }
 
 #[derive(Fail, Debug)]
