+++ crates/fs-utils/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "fs-utils"
+version = "0.1.0"
+authors = ["Michael Stewart <mikrostew@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+failure = "0.1.1"
