+++ crates/volta-core/Cargo.toml
@@ -29,6 +29,7 @@ lazy_static = "1.3.0"
 semver = { git = "https://github.com/mikrostew/semver", branch = "new-parser" }
 cmdline_words_parser = "0.0.2"
 reqwest = { version = "0.9.9", features = ["hyper-011"] }
+fs-utils = { path = "../fs-utils" }
 headers-011 = { path = "../headers-011" }
 cfg-if = "0.1"
 winfolder = "0.1"
